3.1

+++ src/main/java/com/techacademy/repository/ReportRepository.java
@@ -2,10 +2,10 @@ package com.techacademy.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.techacademy.entity.Report;
-import com.techacademy.entity.Employee; 
-import java.util.List; 
+import com.techacademy.entity.Employee;
+import java.util.List;
 
-public interface ReportRepository extends JpaRepository<Report, Long> {
+public interface ReportRepository extends JpaRepository<Report, Integer> {
     // 従業員に紐づく日報情報を取得
     List<Report> findByEmployee(Employee employee);
 }
